requirements.txt
fastapi>=0.68.0,<0.116.0
uvicorn>=0.15.0,<0.23.0
psycopg2-binary>=2.9.3,<3.0.0
requests>=2.27.1,<3.0.0
pydantic>=1.8.0,<3.0.0
python-jose>=3.3.0,<4.0.0
spacy>=3.4.0,<4.0.0
python-multipart>=0.0.5
passlib>=1.7.4
bcrypt>=3.2.0
jinja2>=3.0.0
python-dotenv>=0.19.0
email-validator>=1.1.3

# For development
pytest>=7.0.0
black>=22.1.0
flake8>=4.0.1

# py -m spacy download fr_core_news_md
